renamed internal types to flow types
