+++ src/flowTypes/founddation/entityState.js
