+++ src/index.js
@@ -2,7 +2,7 @@ import createActions, { actionTypes } from './actions'
 import { cachePolicies } from './cachePolicies'
 import callApi from './callApi'
 import createConnect from './components'
-import type { ApiTypeMap } from './internalTypes'
+import type { ApiTypeMap } from './flowTypes'
 import createReducer from './reducers'
 import createSaga from './sagas'
 import * as typeUtils from './types'
