+++ src/types/getIdAttribute.js
@@ -1,5 +1,5 @@
 // @flow
-import { type ApiTypeMap, type EntityType } from '../internalTypes'
+import { type ApiTypeMap, type EntityType } from '../flowTypes'
 
 // Temporary workaround: normalizr 3.x has dropped getIdAttribute
 const getIdAttribute = (types: ApiTypeMap, entityType: EntityType) => 'id'
